Add new locking system option
The new locking system is using pthread_rwlock for supporting
wider OSs in one logic.
